Just using DIS instead of v6 and v7 (protocol version is a selectable enumeration)
