+++ src/dis/common/angular_velocity_vector.rs
