+++ src/dis/common/velocity_vector.rs
