+++ src/dis/entity_information/general_appearance.rs
