+++ src/dis/simulation_management/action_request_pdu.rs
@@ -1,38 +1,38 @@
-use crate::dis6::common::{
+use bytes::{Buf, BufMut, BytesMut};
+use std::any::Any;
+
+use crate::dis::common::{
     dis_error::DISError,
     entity_id::EntityId,
     pdu::Pdu,
     pdu_header::{PduHeader, PduType, ProtocolFamily},
 };
-use bytes::{Buf, BufMut, BytesMut};
-use serde::{Deserialize, Serialize};
-use std::any::Any;
 
 #[derive(Copy, Clone, Debug)]
-pub struct ActionResponsePdu {
+pub struct ActionRequestPdu {
     pub pdu_header: PduHeader,
     pub originating_entity_id: EntityId,
     pub receiving_entity_id: EntityId,
     pub request_id: u32,
-    pub request_status: RequestStatus,
+    pub action_id: u32,
     pub number_of_fixed_datum_records: u32,
     pub number_of_variable_datum_records: u32,
     pub fixed_datum_records: u64,
     pub variable_datum_records: u64,
 }
 
-impl ActionResponsePdu {
+impl ActionRequestPdu {
     pub fn default() -> Self {
-        ActionResponsePdu {
+        ActionRequestPdu {
             pdu_header: PduHeader::default(
-                PduType::ActionResponse,
+                PduType::ActionRequest,
                 ProtocolFamily::SimulationManagement,
                 56,
             ),
             originating_entity_id: EntityId::default(1),
             receiving_entity_id: EntityId::default(2),
             request_id: 0,
-            request_status: RequestStatus::Other,
+            action_id: 0,
             number_of_fixed_datum_records: 0,
             number_of_variable_datum_records: 0,
             fixed_datum_records: 0,
@@ -41,13 +41,13 @@ impl ActionResponsePdu {
     }
 }
 
-impl Pdu for ActionResponsePdu {
+impl Pdu for ActionRequestPdu {
     fn serialize(&self, buf: &mut BytesMut) {
         self.pdu_header.serialize(buf);
         self.originating_entity_id.serialize(buf);
         self.receiving_entity_id.serialize(buf);
         buf.put_u32(self.request_id as u32);
-        buf.put_u32(self.request_status as u32);
+        buf.put_u32(self.action_id as u32);
         buf.put_u32(self.number_of_fixed_datum_records);
         buf.put_u32(self.number_of_variable_datum_records);
         buf.put_u64(self.fixed_datum_records);
@@ -59,11 +59,11 @@ impl Pdu for ActionResponsePdu {
         Self: Sized,
     {
         let pdu_header = PduHeader::decode(&mut buffer);
-        if pdu_header.pdu_type == PduType::ActionResponse {
+        if pdu_header.pdu_type == PduType::ActionRequest {
             let originating_entity_id = EntityId::decode(&mut buffer);
             let receiving_entity_id = EntityId::decode(&mut buffer);
             let request_id = buffer.get_u32();
-            let request_status = RequestStatus::decode(&mut buffer);
+            let action_id = buffer.get_u32();
             let number_of_fixed_datum_records = buffer.get_u32();
             let number_of_variable_datum_records = buffer.get_u32();
             let mut fixed_datum_records: u64 = 0;
@@ -75,12 +75,12 @@ impl Pdu for ActionResponsePdu {
                 variable_datum_records += buffer.get_u64();
             }
 
-            return Ok(ActionResponsePdu {
+            return Ok(ActionRequestPdu {
                 pdu_header,
                 originating_entity_id,
                 receiving_entity_id,
                 request_id,
-                request_status,
+                action_id,
                 number_of_fixed_datum_records,
                 number_of_variable_datum_records,
                 fixed_datum_records,
@@ -105,7 +105,7 @@ impl Pdu for ActionResponsePdu {
         let originating_entity_id = EntityId::decode(&mut buffer);
         let receiving_entity_id = EntityId::decode(&mut buffer);
         let request_id = buffer.get_u32();
-        let request_status = RequestStatus::decode(&mut buffer);
+        let action_id = buffer.get_u32();
         let number_of_fixed_datum_records = buffer.get_u32();
         let number_of_variable_datum_records = buffer.get_u32();
         let mut fixed_datum_records: u64 = 0;
@@ -117,12 +117,12 @@ impl Pdu for ActionResponsePdu {
             variable_datum_records += buffer.get_u64();
         }
 
-        return Ok(ActionResponsePdu {
+        return Ok(ActionRequestPdu {
             pdu_header,
             originating_entity_id,
             receiving_entity_id,
             request_id,
-            request_status,
+            action_id,
             number_of_fixed_datum_records,
             number_of_variable_datum_records,
             fixed_datum_records,
@@ -131,56 +131,16 @@ impl Pdu for ActionResponsePdu {
     }
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub enum RequestStatus {
-    Other = 0,
-    Pending = 1,
-    Executing = 2,
-    PartiallyComplete = 3,
-    Complete = 4,
-    RequestRejected = 5,
-    RetransmitRequestNow = 6,
-    RetransmitRequestLater = 7,
-    InvalidTimeParameters = 8,
-    SimulationTimeExceeded = 9,
-    RequestDone = 10,
-    TACCSFLOSReplyTypeOne = 100,
-    TACCSFLOSReplyTypeTwo = 101,
-    JoinExerciseRequestRejected = 201,
-}
-
-impl RequestStatus {
-    pub fn decode(buf: &mut BytesMut) -> RequestStatus {
-        match buf.get_u32() {
-            0 => RequestStatus::Other,
-            1 => RequestStatus::Pending,
-            2 => RequestStatus::Executing,
-            3 => RequestStatus::PartiallyComplete,
-            4 => RequestStatus::Complete,
-            5 => RequestStatus::RequestRejected,
-            6 => RequestStatus::RetransmitRequestNow,
-            7 => RequestStatus::RetransmitRequestLater,
-            8 => RequestStatus::InvalidTimeParameters,
-            9 => RequestStatus::SimulationTimeExceeded,
-            10 => RequestStatus::RequestDone,
-            100 => RequestStatus::TACCSFLOSReplyTypeOne,
-            101 => RequestStatus::TACCSFLOSReplyTypeTwo,
-            201 => RequestStatus::JoinExerciseRequestRejected,
-            _ => RequestStatus::Other,
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::ActionResponsePdu;
-    use crate::dis6::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
+    use super::ActionRequestPdu;
+    use crate::dis::common::pdu_header::{PduHeader, PduType, ProtocolFamily};
 
     #[test]
     fn create_header() {
-        let action_request_pdu = ActionResponsePdu::default();
+        let action_request_pdu = ActionRequestPdu::default();
         let pdu_header = PduHeader::default(
-            PduType::ActionResponse,
+            PduType::ActionRequest,
             ProtocolFamily::SimulationManagement,
             448 / 8,
         );
